requirements.txt
# === Core database connectors ===
mysql-connector-python>=9.0.0
pymongo>=4.7.0

# === Data analysis ===
pandas>=2.2.2
numpy>=1.26.4

# === Data collection / scraping ===
requests>=2.32.3
beautifulsoup4>=4.12.3
yfinance>=0.2.43

# === Machine learning / feature analysis ===
scikit-learn>=1.5.2

# === Utilities ===
python-dotenv>=1.0.1
schedule>=1.2.1
tqdm>=4.66.5